Refactor: Padronização dos arquivos
